memref-test: munmap correct size
We don't check the result of munmap because it is in the cleanup path,
but it was failing even in the correct test path due to an incorrect
size. This failure would cause memory sent to the receiver to not be
relinquished to NS at the end of the test.

Bug: 178613503
Test: Run tipc-test -t send-fd with linux logging, see handle return
Change-Id: I9b2272ae6b98cbeeb69262058737587a595a6ceb