[hwcrypto] Add generic support for opaque key handles
An opaque key handle uniquely identifies and gives adccess to a keyslot
for a particular client without exposing the key material to that
client. Given a handle, a delegated process (e.g. hwkey) will be able
to, if allowed, retrieve the key material.

Add structures and handler for opaque key handles. Key retrieval for
hwaes will follow in a separate change.

Bug: 185483587
Test: build.py --test com.android.trusty.hwcrypto.test
Change-Id: I8e0fcfc24bb8f03b9075d06418cf9132ebfe4884