[hwkey] Move test keyslots into main keyslot array
We will need to search the keyslots array to handle opaque keys. If the
test keyslots are in a separate array we would need further duplication
for handling opaque keys in both keyslot arrays.

Move the test keyslots into the regular keyslot array and delete the
special dispatch logic when building with hwcrypto unittests enabled.
These test keyslots will only be present when the define for hwcrypto
unittests is set.

Bug: 185483587
Test: com.android.trusty.hwcrypto.test
Change-Id: I737f79b6a08c42470a72c291b394f6ec90730b8a