[hwkey] Add derived key support for keyslots
Some devices may store a shared encryption key in hardware. However, we
do not want to alllow multiple clients to directly use this key, as they
would then be able to decrypt each other's data. To solve this, we want
to be able to derive unique, client-specific keys from a shared
encryption key.

Adds generic support for embedding an encrypted key for a key slot and
decrypting it when retrieving the key slot. This facilitates support for
providing unique keys to different clients that are all derived from the
same underlying key.

To use this capability for key derivation from a common shared key, the
encrypting entity should generate a unique, random key for a particular
client, then encrypt that unique key using the common shared key
resulting in a wrapped, client-specific key. This wrapped key can then
be safely embedded in the hwkey service and will only be accessible
using the shared key.

Bug: 188711856
Test: com.android.trusty.hwcrypto.test

Change-Id: I16c79bc56fc0b2275c0c49cee928288eb837a866