[hwkey] Restrict usage of app loading keys based on unlock state
Check the system_state service for the current state of
SYSTEM_STATE_FLAG_APP_LOADING_UNLOCKED and restrict the use of app
loader keys based on this state.

To test this we want to ensure that only unlocked devices can load apps
signed using the slot 1 signing key. Adds an otherwise empty loadable
app service that is signed with the slot 1 key. Tests that this app is
loadable if the unlocked system state flag is true and loading results
in a verification error otherwise.

Bug: 189260268
Test: change STATIC_SYSTEM_STATE_FLAG_APP_LOADING_UNLOCKED to 0, run
      com.android.appmgmt-unittest.appmgr
Change-Id: Ib35a5ef4d1dec9d52f6dd838dbda6700e948401c