[hwcrypto] Add support for combining opaque and derived keys
We may want to use opaque handles to referred to derived keys. If we
nest a derived key inside an opaque key, we can no longer rely on the
slot->priv opaque pointer containing a hwkey_derived_keyslot_data
structure, since it will instead be a hwkey_opaque_handle_data in the
nested case. Instead, retriever callbacks now take a pointer to the
handle data they are part of, and are required to dereference this
handle to get to the opaque private data if needed.

Adds a test with an opaque + derived keyslot.

Test: com.android.trusty.hwcrypto.test
Bug: 191402516
Change-Id: Ib474e4387d34496fef5482e9e2c8125442ad22e8